[LEOP-318] Extract SSR feature (#136)
* [LEOP-318] Extract SSR feature

* fix path

* update brs version and CHANGELOG

* revert version to 10.0.0
