+++ packages/react-scripts/backpack-addons/ssr/isSsr.js
