+++ packages/react-scripts/scripts/build.js
@@ -45,7 +45,7 @@ const printHostingInstructions = require('react-dev-utils/printHostingInstructio
 const FileSizeReporter = require('react-dev-utils/FileSizeReporter');
 const printBuildError = require('react-dev-utils/printBuildError');
 
-const isSsr = require('../scripts/utils/isSsr');
+const isSsr = require('../backpack-addons/ssr/isSsr');
 
 const measureFileSizesBeforeBuild =
   FileSizeReporter.measureFileSizesBeforeBuild;
